Move sample generation from README.md
The problem with building `//...` is that bazel downloads all the build toolchains, even though only one is useful for the host.

By moving the sample generation in //sample, anyone can run `bazel build //sample/...` and only download what is needed to produce the README file in all formats.

Also, exports PANDOC_FORMATS so that the rule and the sample use the same set of formats.